Version 1.6 - Rewards and points, functional settings, spawn handling
Add rewards and points system. Give configurable reward point amounts to
a player when they:
- join the server for the first time
- vote for a server supported by Votifier
- claim their daily reward

Add functionality to settings. Toggling a setting on or off will now be
used to check against the specific task.

Add player respawn and player join events to handle a player location
when they respawn after dying and join the server for the first time.

Add vote and donate config section to customise the /vote and /donate
messages.

Remove a number of unused classes.